moved navbar.js to components/layout
