+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 import './App.css';
-import Navbar from './Components/Navbar';
+import Navbar from './Components/layout/Navbar';
 
 class App extends Component {
   render() {
